Code clean up using pre-commit

+++ tele_bot.py
@@ -1,29 +1,28 @@
-import telebot.async_telebot as atb
-import pandas as pd
 import asyncio
-import battler as bt
-from pathlib import Path
 import os
+from pathlib import Path
+
+import battler as bt
+from dsl.interpreter import parse_file
+import pandas as pd
+import telebot.async_telebot as atb
 
 b = bt.Battler(game_cls=bt.Kalah, game_run="play_alpha_beta")
 
-dummy = Path('mail_test/fedor_novikov.py')
+dummy = Path("mail_test/fedor_novikov.py")
 
-bot = atb.AsyncTeleBot('5998949800:AAGj_6DfBQEbQLe-tC5B78ZdBfy3IG4ALko');
-table = pd.DataFrame({
-    'name': [],
-    'code': [],
-    'score': []
-})
-table.index.name = 'id'
-save_dir = 'mail_saved'
+bot = atb.AsyncTeleBot("5998949800:AAGj_6DfBQEbQLe-tC5B78ZdBfy3IG4ALko")
+table = pd.DataFrame({"name": [], "code": [], "score": []})
+table.index.name = "id"
+save_dir = "mail_saved"
 
 
-
-@bot.message_handler(content_types=['text'])
+@bot.message_handler(content_types=["text"])
 async def get_text_messages(message):
     if message.text == "/start":
-        await bot.send_message(message.from_user.id, "Введите /register ФАМИЛИЯ_ИМЯ_ГРУППА для регистрации")
+        await bot.send_message(
+            message.from_user.id, "Введите /register ФАМИЛИЯ_ИМЯ_ГРУППА для регистрации"
+        )
     elif message.text[0:9] == "/register":
         await register(message)
     elif message.text == "/start_bt":
@@ -42,50 +41,54 @@ async def get_text_messages(message):
         await bot.send_message(message.from_user.id, "Я тебя не понимаю.")
 
 
-@bot.message_handler(content_types=['document'])
+@bot.message_handler(content_types=["document"])
 async def get_doc_messages(message):
     file_name = message.document.file_name
     file_id = message.document.file_name
     file_id_info = await bot.get_file(message.document.file_id)
     downloaded_file = await bot.download_file(file_id_info.file_path)
-    src = str(message.from_user.id) + '_test' + '.py'
-    old_file = str(message.from_user.id) + '.py'
+    src = str(message.from_user.id) + "_test" + ".py"
+    old_file = str(message.from_user.id) + ".py"
     save_path = save_dir + "/" + src
-
-    with open(save_path, 'wb') as new_file:
-        new_file.write(downloaded_file)
-
-    res = await b.run_dummy(Path(save_path), dummy, func_name='func')
-    if isinstance(res, str):
-        await bot.send_message(message.from_user.id, res)
-        os.remove(save_path)
+    save_path = parse_file(Path(save_path))
+    if save_path:
+        with open(save_path, "wb") as new_file:
+            new_file.write(downloaded_file)
+        res = await b.run_dummy(Path(save_path), dummy, func_name="func")
+        if isinstance(res, str):
+            await bot.send_message(message.from_user.id, res)
+            os.remove(save_path)
+        else:
+            if os.path.exists(save_dir + "/" + old_file):
+                print("Remove old file")
+                os.remove(save_dir + "/" + old_file)
+            os.rename(save_path, save_dir + "/" + old_file)
+            await bot.send_message(message.from_user.id, "Сохранил")
+            table.loc[message.from_user.id, "code"] = Path(save_dir + "/" + old_file)
+            print("Add FILE " + str(message.from_user.id))
     else:
-        if os.path.exists(save_dir + "/" + old_file):
-            print('Remove old file')
-            os.remove(save_dir + "/" + old_file)
-        os.rename(save_path, save_dir + "/" + old_file)
-        await bot.send_message(message.from_user.id, 'Сохранил')
-        table.loc[message.from_user.id, 'code'] = Path(save_dir + "/" + old_file)
-        print('Add FILE ' + str(message.from_user.id))
+        await bot.send_message(message.from_user.id, "File has incorrect extension.")
 
 
 async def register(message):
-    if (len(message.text)<11):
-        await bot.send_message(message.from_user.id, 'Некорректное имя')
+    if len(message.text) < 11:
+        await bot.send_message(message.from_user.id, "Некорректное имя")
         return
-    table.loc[message.from_user.id, 'name'] = message.text[9:len(message.text)]
-    await bot.send_message(message.from_user.id, "Можешь загрузить свой файл в любое время")
-    print('Add people ' + str(message.from_user.id))
+    table.loc[message.from_user.id, "name"] = message.text[9 : len(message.text)]
+    await bot.send_message(
+        message.from_user.id, "Можешь загрузить свой файл в любое время"
+    )
+    print("Add people " + str(message.from_user.id))
 
 
 async def save_table():
-    table.to_excel('dsl_table.xlsx')
-    print('Saving table')
+    table.to_excel("dsl_table.xlsx")
+    print("Saving table")
 
 
 async def load_table():
-    temp_table = pd.read_excel('dsl_table.xlsx', index_col='id')
-    print('Load table')
+    temp_table = pd.read_excel("dsl_table.xlsx", index_col="id")
+    print("Load table")
     return temp_table
 
 
@@ -94,7 +97,7 @@ async def print_table():
 
 
 async def send_text_mes(mes):
-    print('Send all msg')
+    print("Send all msg")
     for user in table.index:
         await bot.send_message(user, mes)
 
@@ -104,21 +107,23 @@ def run_test():
 
 
 async def send_result(res):
-    print('отпрвка')
+    print("отпрвка")
     for item in res.items():
-        table.loc[int(item[0]), 'score'] = item[1]
-        await bot.send_message(item[0], 'Ваш результат в последнем турнире: %.1f' % item[1])
-
+        table.loc[int(item[0]), "score"] = item[1]
+        await bot.send_message(
+            item[0], "Ваш результат в последнем турнире: %.1f" % item[1]
+        )
 
 
 async def start_battle():
-    await send_text_mes('Раунд начинается, ваше последнее решение примет участие в турнире')
+    await send_text_mes(
+        "Раунд начинается, ваше последнее решение примет участие в турнире"
+    )
     b.check_contestants(Path(save_dir), func_name="func")
     await b.run_tournament(n_workers=4, timeout=2.5)
-    res = b.form_results();
+    res = b.form_results()
     await send_result(res)
     b.save_results(Path("result.json"))
-
 
 
 if __name__ == "__main__":
